Refactor api file structure
