+++ src/api/newGoal/const_vars.js
@@ -1,5 +1,3 @@
-
-
 //-------- DEFAULT IMAGE SIZE --------------
 const DFLT_IMG_SIZE={
     width:100,
